create separate_cases_file test data
