+++ testrunner/testdata/practice/separate_cases_file/.meta/config.json
@@ -1,0 +1,23 @@
+{
+  "blurb": "...",
+  "authors": [
+    "..."
+  ],
+  "contributors": [
+    "..."
+  ],
+  "files": {
+    "solution": [
+      "conditionals.go"
+    ],
+    "test": [
+      "conditionals_test.go"
+    ],
+    "example": [
+      ".meta/example.go"
+    ]
+  },
+  "custom": {
+    "taskIdsEnabled": false
+  }
+}
