+++ testrunner/testdata/practice/separate_cases_file/go.mod
@@ -1,0 +1,3 @@
+module conditionals
+
+go 1.18
